requirements.txt
flask
requests
fpdf
python-docx
